# swagger.json
{
   "swagger": "2.0",
   "info": {
      "title": "AutoMart API Documentation",
      "description": "This is the documentation for Auto Mart API, an online marketplace where users can buy and sell cars of diverse makes.",
      "version": "1.1.0"
   },
   "host": "auto-mart-api.herokuapp.com",
   "basePath": "/api/v1",
   "schemes": [
      "https"
   ],
...
   "definitions": {
      "User": {
         "type": "object",
         "properties": {
            "_id": {
               "type": "string",
               "description": "The user ID."
            },
            "first_name": {
               "type": "string",
               "description": "The user first name."
            },
            "last_name": {
               "type": "string",
               "description": "The user last name."
            },
            "profile_pic": {
               "properties": {
                  "items": {
                     "type": "array",
                     "items": {}
                  }
               },
               "description": "The user profile picture details containing the image url and id."
            },
            "phone": {
               "type": "string",
               "description": "The user phone number."
            },
            "address": {
               "type": "string",
               "description": "The user address."
            },
            "email": {
               "type": "string",
               "description": "The user email."
            },
            "password": {
               "type": "string",
               "description": "The user password."
            },
            "joined": {
               "type": "string",
               "format": "date-time",
               "description": "The user registration date."
            }
         }
      },
...
   },
   "securityDefinitions": {
      "BearerAuth": {
         "type": "apiKey",
         "in": "header",
         "name": "Authorization"
      }
   },
...
   "paths": {
      "/auth/signup": {
         "post": {
            "description": "Registers a new user.",
            "tags": [
               "auth"
            ],
            "consumes": [
               "multipart/form-data"
            ],
            "parameters": [
               {
                  "in": "formData",
                  "name": "first_name",
                  "type": "string",
                  "required": true,
                  "description": "User first name"
               },
               {
                  "in": "formData",
                  "name": "last_name",
                  "type": "string",
                  "required": true,
                  "description": "User last name"
               },
               {
                  "in": "formData",
                  "name": "profile_pic",
                  "type": "file",
                  "required": true,
                  "description": "User profile picture."
               },
               {
                  "in": "formData",
                  "name": "phone",
                  "type": "string",
                  "required": true,
                  "description": "User phone number"
               },
               {
                  "in": "formData",
                  "name": "address",
                  "type": "string",
                  "required": true,
                  "description": "User address."
               },
               {
                  "in": "formData",
                  "name": "email",
                  "type": "string",
                  "required": true,
                  "description": "User email."
               },
               {
                  "in": "formData",
                  "name": "password",
                  "type": "string",
                  "required": true,
                  "description": "User password."
               },
               {
                  "in": "formData",
                  "name": "confirmPass",
                  "type": "string",
                  "required": true,
                  "description": "Retype password for confirmation."
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "schema": {
                     "$ref": "#/definitions/User"
                  }
               }
            }
         }
      },
      "/auth/signin": {
         "post": {
            "description": "Login a user.",
            "tags": [
               "auth"
            ],
            "consumes": [
               "application/x-www-form-urlencoded"
            ],
            "parameters": [
               {
                  "in": "body",
                  "name": "User Login",
                  "description": "User login details.",
                  "schema": {
                     "type": "object",
                     "required": [
                        "email",
                        "password"
                     ],
                     "properties": {
                        "email": {
                           "type": "string"
                        },
                        "password": {
                           "type": "string"
                        }
                     }
                  }
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "schema": {
                     "$ref": "#/definitions/User"
                  }
               }
            }
         }
      },
      "/auth/delete_user/{user_id}/": {
         "delete": {
            "security": [
               {
                  "BearerAuth": []
               }
            ],
            "description": "Deletes a user.",
            "tags": [
               "auth"
            ],
            "parameters": [
               {
                  "name": "user_id",
                  "in": "path",
                  "description": "5da25a1d0fa5bf00171f96d2",
                  "type": "string",
                  "required": true
               }
            ],
            "produces": [
               "application/json"
            ],
            "responses": {
               "200": {
                  "description": "OK"
               }
            }
         }
      },
...
   }
}